Project 11-Chat-Over-RAG/requirements.txt
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .
sentence-transformers
pinecone